pyproject.toml
[project]
name = "mujoco_depth_pointcloud"
version = "0.1.0"
description = "MuJoCo-based pointcloud generator with RGB-D simulation"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.2.6",
    "opencv-python>=4.12.0.88",
    "mujoco>=3.3.5",
    "glfw",
    "matplotlib>=3.5.0",
    "rerun-sdk>=0.20.0",
]
